#770: Extract hook definition setup from FluentInjector
