+++ fluentlenium-core/src/main/java/org/fluentlenium/core/inject/FluentInjector.java
@@ -1,11 +1,8 @@
 package org.fluentlenium.core.inject;
 
-import static org.fluentlenium.core.inject.InjectionAnnotationSupport.isAnnotationTypeHook;
-import static org.fluentlenium.core.inject.InjectionAnnotationSupport.isAnnotationTypeHookOptions;
 import static org.fluentlenium.core.inject.InjectionAnnotationSupport.isContainer;
 import static org.fluentlenium.core.inject.InjectionAnnotationSupport.isParent;
 
-import org.apache.commons.lang3.ArrayUtils;
 import org.fluentlenium.core.FluentContainer;
 import org.fluentlenium.core.FluentControl;
 import org.fluentlenium.core.components.ComponentsManager;
@@ -14,25 +11,16 @@ import org.fluentlenium.core.components.LazyComponentsListener;
 import org.fluentlenium.core.events.ContainerAnnotationsEventsRegistry;
 import org.fluentlenium.core.events.EventsRegistry;
 import org.fluentlenium.core.hook.DefaultHookChainBuilder;
-import org.fluentlenium.core.hook.FluentHook;
-import org.fluentlenium.core.hook.Hook;
-import org.fluentlenium.core.hook.HookControlImpl;
 import org.fluentlenium.core.hook.HookDefinition;
-import org.fluentlenium.core.hook.HookOptions;
-import org.fluentlenium.core.hook.NoHook;
 import org.fluentlenium.core.proxy.LocatorProxies;
 import org.fluentlenium.utils.ReflectionUtils;
 import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.pagefactory.ElementLocator;
 
-import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.IdentityHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.function.Consumer;
@@ -40,7 +28,6 @@ import java.util.function.Consumer;
 /**
  * Handle injection of element proxies, @Page objects and @FindBy.
  */
-@SuppressWarnings("PMD.GodClass")
 public class FluentInjector implements FluentInjectControl {
 
     private final Map<Class, Object> containerInstances = new IdentityHashMap<>();
@@ -55,6 +42,7 @@ public class FluentInjector implements FluentInjectControl {
 
     private final FluentElementInjectionSupportValidator injectionSupportValidator;
     private final FluentInjectFieldInitializer fieldInitializer;
+    private final FluentInjectHookDefinitionAdder hookDefinitionAdder = new FluentInjectHookDefinitionAdder();
 
     /**
      * Creates a new injector.
@@ -154,7 +142,7 @@ public class FluentInjector implements FluentInjectControl {
         }
 
         for (Class cls = container.getClass(); isClassSupported(cls); cls = cls.getSuperclass()) {
-            addHookDefinitions(cls.getDeclaredAnnotations(), containerContext.getHookDefinitions());
+            hookDefinitionAdder.addHookDefinitions(cls.getDeclaredAnnotations(), containerContext.getHookDefinitions());
         }
     }
 
@@ -200,7 +188,7 @@ public class FluentInjector implements FluentInjectControl {
         forAllDeclaredFieldsInHierarchyOf(container, field -> {
             if (injectionSupportValidator.isSupported(container, field)) {
                 ArrayList<HookDefinition<?>> fieldHookDefinitions = new ArrayList<>(containerContext.getHookDefinitions());
-                addHookDefinitions(field.getAnnotations(), fieldHookDefinitions);
+                hookDefinitionAdder.addHookDefinitions(field.getAnnotations(), fieldHookDefinitions);
                 InjectionElementLocatorFactory locatorFactory = new InjectionElementLocatorFactory(searchContext);
                 InjectionElementLocator locator = locatorFactory.createLocator(field);
                 if (locator != null) {
@@ -242,84 +230,5 @@ public class FluentInjector implements FluentInjectControl {
             return lazyComponents.isLazy() && !lazyComponents.isLazyInitialized();
         }
         return false;
-    }
-
-    private Hook getHookAnnotation(Annotation annotation) {
-        if (annotation instanceof Hook) {
-            return (Hook) annotation;
-        } else if (isAnnotationTypeHook(annotation)) {
-            return annotation.annotationType().getAnnotation(Hook.class);
-        }
-        return null;
-    }
-
-    private HookOptions getHookOptionsAnnotation(Annotation annotation) {
-        if (annotation instanceof HookOptions) {
-            return (HookOptions) annotation;
-        } else if (isAnnotationTypeHookOptions(annotation)) {
-            return annotation.annotationType().getAnnotation(HookOptions.class);
-        }
-        return null;
-    }
-
-    private void addHookDefinitions(Annotation[] annotations, List<HookDefinition<?>> hookDefinitions) {
-        Hook currentHookAnnotation = null;
-        HookOptions currentHookOptionAnnotation = null;
-
-        Annotation currentAnnotation = null;
-        for (Annotation annotation : annotations) {
-            applyNoHook(hookDefinitions, annotation);
-
-            Hook hookAnnotation = getHookAnnotation(annotation);
-            if (hookAnnotation != null) {
-                currentAnnotation = annotation;
-                if (currentHookAnnotation != null) {
-                    hookDefinitions.add(buildHookDefinition(currentHookAnnotation, currentHookOptionAnnotation, currentAnnotation));
-                    currentHookOptionAnnotation = null;
-                }
-                currentHookAnnotation = hookAnnotation;
-            }
-            HookOptions hookOptionsAnnotation = getHookOptionsAnnotation(annotation);
-            if (hookOptionsAnnotation != null) {
-                if (currentHookOptionAnnotation != null) {
-                    throw new FluentInjectException("Unexpected @HookOptions annotation. @Hook is missing.");
-                }
-                currentHookOptionAnnotation = hookOptionsAnnotation;
-            }
-        }
-
-        if (currentHookAnnotation != null) {
-            hookDefinitions.add(buildHookDefinition(currentHookAnnotation, currentHookOptionAnnotation, currentAnnotation));
-        }
-    }
-
-    private void applyNoHook(List<HookDefinition<?>> hookDefinitions, Annotation annotation) {
-        if (annotation instanceof NoHook) {
-            Hook[] value = ((NoHook) annotation).value();
-            if (ArrayUtils.isEmpty(value)) {
-                hookDefinitions.clear();
-            } else {
-                HookControlImpl
-                        .removeHooksFromDefinitions(hookDefinitions, Arrays.stream(value).map(Hook::value).toArray(Class[]::new));
-            }
-        }
-    }
-
-    private <T> HookDefinition<T> buildHookDefinition(Hook hookAnnotation, HookOptions hookOptionsAnnotation,
-            Annotation currentAnnotation) {
-        Class<? extends T> hookOptionsClass =
-                hookOptionsAnnotation == null ? null : (Class<? extends T>) hookOptionsAnnotation.value();
-        T fluentHookOptions = null;
-        if (hookOptionsClass != null) {
-            try {
-                fluentHookOptions = ReflectionUtils.newInstanceOptionalArgs(hookOptionsClass, currentAnnotation);
-            } catch (NoSuchMethodException e) {
-                throw new FluentInjectException("@HookOption class has no valid constructor", e);
-            } catch (IllegalAccessException | InvocationTargetException | InstantiationException e) {
-                throw new FluentInjectException("Can't create @HookOption class instance", e);
-            }
-        }
-        Class<? extends FluentHook<T>> hookClass = (Class<? extends FluentHook<T>>) hookAnnotation.value();
-        return fluentHookOptions == null ? new HookDefinition<>(hookClass) : new HookDefinition<>(hookClass, fluentHookOptions);
     }
 }
